@@ -23,7 +23,7 @@ class local extends \phpbb\avatar\driver\driver
 	*/
 	public function get_data($row)
 	{
-		$root_path = (defined('PHPBB_USE_BOARD_URL_PATH') && PHPBB_USE_BOARD_URL_PATH) ? generate_board_url() . '/' : $this->path_helper->get_web_root_path();
+		$root_path = $this->path_helper->get_web_root_path();
 
 		return array(
 			'src' => $root_path . $this->config['avatar_gallery_path'] . '/' . $row['avatar'],
