@@ -314,11 +314,19 @@ class factory implements driver_interface
 	}
 
 	/**
-	* {@inheritdoc}
-	*/
+	 * {@inheritdoc}
+	 */
 	public function sql_nextid()
 	{
-		return $this->get_driver()->sql_nextid();
+		return $this->get_driver()->sql_last_inserted_id();
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	public function sql_last_inserted_id()
+	{
+		return $this->get_driver()->sql_last_inserted_id();
 	}
 
 	/**
@@ -463,5 +471,13 @@ class factory implements driver_interface
 	public function sql_quote($msg)
 	{
 		return $this->get_driver()->sql_quote($msg);
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public function clean_query_id($query_id)
+	{
+		return $this->get_driver()->clean_query_id($query_id);
 	}
 }
