@@ -13,6 +13,8 @@
 
 namespace phpbb\install\controller;
 
+use phpbb\path_helper;
+
 class installer_index
 {
 	/**
@@ -24,6 +26,9 @@ class installer_index
 	 * @var \phpbb\language\language
 	 */
 	protected $language;
+
+	/** @var path_helper */
+	protected $path_helper;
 
 	/**
 	 * @var \phpbb\template\template
@@ -40,13 +45,15 @@ class installer_index
 	 *
 	 * @param helper 					$helper
 	 * @param \phpbb\language\language	$language
+	 * @param path_helper				$path_helper
 	 * @param \phpbb\template\template	$template
 	 * @param string					$phpbb_root_path
 	 */
-	public function __construct(helper $helper, \phpbb\language\language $language, \phpbb\template\template $template, $phpbb_root_path)
+	public function __construct(helper $helper, \phpbb\language\language $language, path_helper $path_helper, \phpbb\template\template $template, $phpbb_root_path)
 	{
 		$this->helper = $helper;
 		$this->language = $language;
+		$this->path_helper = $path_helper;
 		$this->template = $template;
 		$this->phpbb_root_path = $phpbb_root_path;
 	}
@@ -59,7 +66,8 @@ class installer_index
 		{
 			case "intro":
 				$title = $this->language->lang('INTRODUCTION_TITLE');
-				$body = $this->language->lang('INTRODUCTION_BODY');
+				$install_docs_path = $this->path_helper->update_web_root_path($this->phpbb_root_path . 'docs/INSTALL.html');
+				$body = $this->language->lang('INTRODUCTION_BODY', $install_docs_path);
 			break;
 			case "support":
 				$title = $this->language->lang('SUPPORT_TITLE');
